Added section in profile page
