+++ ui-sample/angular-ui/src/app/components/create/create.component.html
@@ -7,7 +7,7 @@
     <div class="ui raised shadow container segment" style="width: 44rem;">
         <h3 class="center tac">{{resourceService?.frmelmnts?.btn?.onBoard}}</h3>
         <div class="d-flex flex-jc-center mb-30" *ngIf="formFieldProperties">
-            <app-default-template #formData [formFieldProperties]="formFieldProperties">
+            <app-default-template #formData [formFieldProperties]="formFieldProperties" [formInputData]="formInputDta">
             </app-default-template>
         </div>
         <div class="buttons flex-jc-center">
