+++ ui-sample/angular-ui/src/app/components/update/update.component.html
@@ -1,34 +1,27 @@
 <div class="ui container mb-20 mt-20">
-    <div class="ui small breadcrumb">
-        <a class="section" routerLink=''>{{resourceService?.frmelmnts?.btn?.home}}</a>
-        <i class="right chevron icon divider"></i>
-        <a class="section"
-            (click) = "navigateToProfilePage()">{{resourceService?.frmelmnts?.label?.profileInformation}}</a>
-        <i class="right chevron icon divider"></i>
-        <div class="active section">{{resourceService?.frmelmnts?.label?.updateProfile}}</div>
+  <div class="ui small breadcrumb">
+    <a class="section" routerLink=''>{{resourceService?.frmelmnts?.btn?.home}}</a>
+    <i class="right chevron icon divider"></i>
+    <a class="section" (click)="navigateToProfilePage()">{{resourceService?.frmelmnts?.label?.profileInformation}}</a>
+    <i class="right chevron icon divider"></i>
+    <div class="active section">{{resourceService?.frmelmnts?.label?.updateProfile}}</div>
+  </div>
+  <div class="ui raised shadow container segment" style="width: 50rem;">
+    <div class="ui segment" *ngIf="showLoader">
+      <div class="ui active medium inline loader"></div>
     </div>
-    <div class="ui raised shadow container segment" style="width: 50rem;">
-            <div class="ui segment" *ngIf="showLoader">
-                    <div class="ui active medium inline loader"></div>
-                  </div>
-        <h3 class="center tac">{{resourceService?.frmelmnts?.btn?.update}}</h3>
-
-        <div *ngFor="let section of sections">
-            <div class="ui raised shadow container segment mb-30" *ngIf="section.fields.length > 0">
-              <h2 style="color: grey;" *ngIf="section.fields.length > 0">{{section.name | titlecase}} section</h2>
-              <div class="ui grid mb-30">
-                <app-default-template #formData [formFieldProperties]="section.fields">
-                </app-default-template>
-              </div>
-            </div>
-          </div>
-        <!-- <div class="d-flex flex-jc-center mb-30" *ngIf="formFieldProperties">
-            <app-default-template #formData [formFieldProperties]="formFieldProperties">
-            </app-default-template>
-        </div> -->
-        <div class="buttons flex-jc-center">
-            <button class="sb-btn sb-btn-normal primary" (click)="validate()">
-                {{resourceService?.frmelmnts?.btn?.update}} </button>
+    <div *ngFor="let section of sections">
+      <div class="ui header tas" *ngIf="section.fields.length > 0">{{section.name | titlecase}} section</div>
+      <div class="ui segment mb-20" *ngIf="section.fields.length > 0">
+        <div class="row">
+          <app-default-template #formData [formFieldProperties]="section.fields" [formInputData]="formInputData">
+          </app-default-template>
         </div>
+      </div>
     </div>
+    <div class="buttons flex-jc-center">
+      <button class="sb-btn sb-btn-normal primary" (click)="validate()">
+        {{resourceService?.frmelmnts?.btn?.update}} </button>
+    </div>
+  </div>
 </div>
